Full curse build added to repo/prep for manifest automation
Mods in override, manifest.json and mods.txt added to repo, downloading development as a zip will now produce a buildable modpack. I will be implementing an automation workflow with GitHub that will allow a new manifest.json based on the Curse Project IDs located in the mods.txt file to be generated each time a build is pushed to a branch. So to add or remove mods from the next build, add or remove curse project ID's from mods.txt. The manifest will reflect these changes when changes are pushed to that text file.