Forge Essentials version change/codinglib autopull fix
Forge Essentials needs to be downgraded due to an unfixed Ars Magica 2 bug that doesn't occur on this particular version added. Overrides must be used if pulling a version besides latest.

Also move codinglib to the proper "coremod" location to prevent the dialog box to download on pack launch/server launch.