"Glacia" planet definitions and parasite spawns.
The Lost Cities dimension now generates the Atlantis generation type, and is registered in the Solar System as "Glacia" (name random and can be changed)

It is flooded and forced to use "frozenocean" biome, creating a huge ice sheet with the occasional opening into the water.

The four global parasite spawns have the ability to spawn here.

Some other changes;
- Reduced maximum mob count for Venus to 50.
- Removed carrier parasites from the normal spawn system (left in error)
- Added Creepers, Blazes, and Ghasts to the "attack parasites" list.
- Removed Waila configs