Added downloadUrl field to manifest for each mod, to allow server downloads.
- Downgraded WarpDrive to 1.5.27, the latest version is not available on the API yet.
- Downgraded AromaCore and AromaBackup to their stable releases
- Downgraded InventoryTweaks to the stable release
- Downgraded LootTableTweaker, latest version is not available on API (despite being rather old).

Adding the "downloadUrl" field to the manifest bypasses (completely?!?!?!?!) the restrictions placed on the API it would seem, placing this field with the direct API download link of a mod allows serverstarter to download all the mods, including those set to not be downloadable from third party such as RTG.

A few mods had no versions that were available to be downloaded this way (unsure as to why, seems very arbitrary. Many of these mods had Releases, with no alpha or beta builds. These mods will be provided in the server package (all have MIT)